+++ CHANGELOG.md
@@ -31,6 +31,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * The `descendants()` revset function now accepts an optional `depth` argument;
   like the `ancestors()` depth argument, it limits the depth of the set.
 
+* Revset/template aliases now support function overloading.
+  [#2966](https://github.com/martinvonz/jj/issues/2966)
+
 ### Fixed bugs
 
 ## [0.18.0] - 2024-06-05